requirements.txt
asgiref==3.7.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.2.4
django-allauth==0.54.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
phonenumbers==8.13.18
protobuf==4.21.1
psycopg2-binary==2.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.9
